requirements.txt
flask==2.3.3
flask-cors==5.0.1
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
werkzeug==2.3.7
inference-sdk==0.1.6
opencv-python==4.8.0.76
supervision==0.16.0 